refactor(SLB-222): font preloadedingˆ
