+++ packages/ui/package.json
@@ -13,8 +13,8 @@
     "./styles.css": [
       "./build/styles.css"
     ],
-    "./fonts": [
-      "./fonts.js"
+    "./preloaded-fonts": [
+      "./build/preloaded-fonts.js"
     ]
   },
   "typesVersions": {
@@ -22,8 +22,8 @@
       "routes/*": [
         "build/components/Routes/*.d.ts"
       ],
-      "fonts": [
-        "fonts.d.ts"
+      "preloaded-fonts": [
+        "build/preloaded-fonts.d.ts"
       ]
     }
   },
@@ -89,6 +89,7 @@
     "@types/react": "^18.2.46",
     "@types/react-body-classname": "^1.1.10",
     "@types/react-dom": "^18.2.18",
+    "@types/node": "~18.15.3",
     "@vitejs/plugin-react-swc": "^3.5.0",
     "autoprefixer": "^10.4.16",
     "cssnano": "^6.0.3",
